LayoutTests/http/tests/iframe-monitor/iframe-unload-expected.txt

CONSOLE MESSAGE: Frame was unloaded because its network usage exceeded the limit: 10240 bytes, url=http://localhost:8080/iframe-monitor/resources/--eligible--/iframe.html
Test iframe with huge resource usage is unloaded.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelector('iframe[name=frame1]').srcdoc is not ""
PASS successfullyParsed is true

TEST COMPLETE
